+++ Reprise.UnitTests/TestTypes.cs
@@ -27,6 +27,13 @@
         }
     }
 
+    public struct StructService : IServiceConfigurator
+    {
+        public void ConfigureServices(WebApplicationBuilder builder)
+        {
+        }
+    }
+
     [Endpoint]
     public class HelloEndpoint
     {
@@ -62,7 +69,11 @@
     {
         public static string NoAttribute() => "Hello, world!";
 
-        [Map("", "/")]
+        [Get("/")]
+        [Post("/")]
+        public static string MultipleAttributes() => "Hello, world!";
+
+        [Map(new[] { "GET", "" }, "/")]
         public static string EmptyHttpMethod() => "Hello, world!";
 
         [Get("")]