+++ Reprise.UnitTests/WebApplicationBuilderExtensionsTests.cs
@@ -1,4 +1,6 @@
-﻿namespace Reprise.UnitTests
+﻿using Extensions = Reprise.WebApplicationBuilderExtensions;
+
+namespace Reprise.UnitTests
 {
     [UsesVerify]
     public class WebApplicationBuilderExtensionsTests
@@ -15,9 +17,18 @@
         public Task ConfigureServices() => Verify(_Builder.ConfigureServices().Services);
 
         [Fact]
-        public Task ConfigureServices_WithAssembly() => Verify(_Builder.ConfigureServices(typeof(Program).Assembly).Services);
+        public Task ConfigureServices_BuilderNull() => Throws(() => Extensions.ConfigureServices(null!));
 
         [Fact]
-        public Task CreateConfigurator() => Throws(() => WebApplicationBuilderExtensions.CreateConfigurator(typeof(ServiceC)));
+        public Task ConfigureServicesAssembly() => Verify(_Builder.ConfigureServices(typeof(Program).Assembly).Services);
+
+        [Fact]
+        public Task ConfigureServicesAssembly_BuilderNull() => Throws(() => Extensions.ConfigureServices(null!, typeof(Program).Assembly));
+
+        [Fact]
+        public Task ConfigureServicesAssembly_AssemblyNull() => Throws(() => _Builder.ConfigureServices(null!));
+
+        [Fact]
+        public Task CreateConfigurator() => Throws(() => Extensions.CreateConfigurator(typeof(ServiceC)));
     }
 }